refactor: move setup/init Commands' logic onto Bench and dedicated classes
Bench.setup_nginx()/setup_letsencrypt() hold the small orchestrations
directly; BenchInitializer (pilot/core/bench_initializer.py) and
ProductionSetup (pilot/core/production_setup.py) hold the larger
init/production-deploy orchestrations, wired through Bench.initialize()/
setup_production(). All four Commands (InitCommand, SetupNginxCommand,
SetupLetsEncryptCommand, SetupProductionCommand) shrink to CLI wiring.

Also extracted BuildAdminCommand's download/build logic to
pilot/core/admin_frontend.py — BenchInitializer and ProductionSetup
needed it, and importing the Command class from pilot/core would have
been a fresh layering violation.

Repoints setup_nginx_task.py, setup_letsencrypt_task.py,
setup_production_task.py, wizard_setup_task.py, and new_site_task.py
(missed in an earlier step) at Bench/Site directly instead of
instantiating Command classes.
