+++ pilot/core/admin_frontend.py
@@ -1,0 +1,112 @@
+from __future__ import annotations
+
+from collections.abc import Callable
+from pathlib import Path
+
+from pilot.exceptions import BenchError
+
+_ADMIN_RELEASE_URL = "https://github.com/frappe/bench-cli/releases/download/latest-build/admin-frontend.tar.gz"
+# The frontend toolchain (unplugin via frappe-ui/vite) uses import.meta.dirname,
+# which only exists in Node 20.11+. Older Node fails the build with an opaque
+# "paths[0] ... undefined" error, so we check up-front.
+_MIN_NODE = (20, 11)
+
+
+def download_admin_frontend(cli_root: Path) -> bool:
+    """Download and extract the pre-built admin frontend. Returns True on success."""
+    import tempfile
+    import urllib.error
+    import urllib.request
+
+    from pilot.archive import extract_tar_archive
+
+    static_dir = cli_root / "admin" / "backend" / "static"
+    with tempfile.NamedTemporaryFile(suffix=".tar.gz", delete=False) as tmp_file:
+        tmp = Path(tmp_file.name)
+
+    print("Downloading admin frontend from GitHub release...", flush=True)
+    try:
+        urllib.request.urlretrieve(_ADMIN_RELEASE_URL, tmp)
+    except urllib.error.URLError as e:
+        print(f"  Download failed: {e}", flush=True)
+        tmp.unlink(missing_ok=True)
+        return False
+
+    try:
+        static_dir.mkdir(parents=True, exist_ok=True)
+        extract_tar_archive(tmp, static_dir)
+        print("  Admin frontend downloaded successfully.", flush=True)
+        return True
+    except Exception as e:
+        print(f"  Extraction failed: {e}", flush=True)
+        return False
+    finally:
+        tmp.unlink(missing_ok=True)
+
+
+def build_admin_frontend(force_build: bool = False, on_progress: Callable[[str], None] = lambda message: None) -> None:
+    from pilot.loader import cli_root
+    from pilot.utils import run_command
+
+    if not force_build and download_admin_frontend(cli_root()):
+        return
+    if force_build:
+        on_progress("Skipping download, building from source...")
+    else:
+        on_progress("Download failed, building from source...")
+    frontend = _find_frontend()
+    _check_node_version()
+    on_progress(f"Building admin frontend at {frontend}...")
+    if _needs_npm_install(frontend):
+        on_progress("Running npm install...")
+        run_command(["npm", "install"], cwd=frontend, stream_output=True)
+    on_progress("Running npm build")
+    run_command(["npm", "run", "build"], cwd=frontend, stream_output=True)
+    on_progress("\nAdmin frontend rebuilt successfully.")
+
+
+def _find_frontend() -> Path:
+    from pilot.loader import cli_root
+
+    candidate = cli_root() / "admin" / "frontend"
+    if (candidate / "package.json").exists():
+        return candidate
+    raise BenchError("admin/frontend not found. This command requires the bench-cli source directory with admin/frontend/.")
+
+
+def _needs_npm_install(frontend: Path) -> bool:
+    install_state = frontend / "node_modules" / ".package-lock.json"
+    if not install_state.exists():
+        return True
+
+    installed_at = install_state.stat().st_mtime
+    for manifest in (frontend / "package.json", frontend / "package-lock.json"):
+        if manifest.exists() and manifest.stat().st_mtime > installed_at:
+            return True
+    return False
+
+
+def _check_node_version() -> None:
+    import subprocess
+
+    try:
+        output = subprocess.run(
+            ["node", "--version"], capture_output=True, text=True, check=True, timeout=5
+        ).stdout.strip()
+    except (FileNotFoundError, subprocess.CalledProcessError, subprocess.TimeoutExpired) as error:
+        raise BenchError(
+            "Node.js is required to build the admin frontend but was not found. "
+            "Install Node.js >= 20.11, or run `bench build-admin` (without --force) to download the pre-built frontend."
+        ) from error
+    parts = output.lstrip("v").split(".")
+    try:
+        version = (int(parts[0]), int(parts[1]))
+    except (IndexError, ValueError):
+        return  # unparseable — let the build run and surface its own error
+    if version < _MIN_NODE:
+        major, minor = _MIN_NODE
+        raise BenchError(
+            f"Building the admin frontend requires Node.js >= {major}.{minor}, but found {output}. "
+            "Switch to a newer Node (e.g. `nvm use 20`) and retry, or run `bench build-admin` "
+            "without --force to download the pre-built frontend instead."
+        )
