+++ pilot/tasks/jobs/new_site_task.py
@@ -1,5 +1,5 @@
-from pilot.commands.sites.create import NewSiteCommand
 from pilot.core.app import App
+from pilot.core.site import Site
 from pilot.integrations.marketplace import Marketplace
 
 from pilot.tasks.jobs.base_task import BaseTask
@@ -26,12 +26,12 @@ class NewSiteTask(BaseTask):
         self._require_production_privileges()
         self._fetch_missing_apps()
         self._step("create", f"Create site {self.name}")
-        NewSiteCommand(self.bench, self.name, self.apps, self.admin_password, db_type=self.db_type).run()
+        Site.provision(self.bench, self.name, self.apps, self.admin_password, db_type=self.db_type, on_progress=self._report)
         self._step("done")
 
     def _fetch_missing_apps(self) -> None:
         """The new-site wizard offers marketplace apps that may not be cloned onto
-        this bench yet; fetch those (and their dependencies) before NewSiteCommand
+        this bench yet; fetch those (and their dependencies) before Site.provision
         validates the app list."""
         apps_txt = self.bench.sites_path / "apps.txt"
         installed = set(apps_txt.read_text().splitlines()) if apps_txt.exists() else set()
