+++ pilot/tasks/jobs/setup_production_task.py
@@ -1,11 +1,10 @@
-from pilot.commands.setup.production import SetupProductionCommand
 from pilot.tasks.jobs.base_task import BaseTask
 
 
 class SetupProductionTask(BaseTask):
     def run(self) -> None:
         self._step("production", "Set up production")
-        SetupProductionCommand(self.bench).run()
+        self.bench.setup_production(on_progress=self._report)
         self._step("done")
