+++ tests/pilot/commands/test_init_command.py
@@ -1,23 +1,23 @@
-"""InitCommand._provision_or_verify: existing database handling."""
+"""BenchInitializer._provision_or_verify: existing database handling."""
 from __future__ import annotations
 
 from unittest.mock import MagicMock
 
 import pytest
 
-from pilot.commands.bench.initialize import InitCommand
+from pilot.core.bench_initializer import BenchInitializer
 from pilot.exceptions import BenchError
 
 
-def _command() -> InitCommand:
-    return InitCommand(MagicMock())
+def _initializer() -> BenchInitializer:
+    return BenchInitializer(MagicMock())
 
 
 def test_provisions_a_pilot_owned_server() -> None:
     manager = MagicMock()
     manager.config.existing = False
 
-    _command()._provision_or_verify(manager, "MariaDB")
+    _initializer()._provision_or_verify(manager, "MariaDB")
 
     manager.provision.assert_called_once()
     manager.check_credentials.assert_not_called()
@@ -28,7 +28,7 @@ def test_verifies_credentials_for_an_existing_server_without_provisioning() -> N
     manager.config.existing = True
     manager.check_credentials.return_value = True
 
-    _command()._provision_or_verify(manager, "MariaDB")
+    _initializer()._provision_or_verify(manager, "MariaDB")
 
     manager.provision.assert_not_called()
     manager.check_credentials.assert_called_once()
@@ -43,6 +43,6 @@ def test_raises_when_existing_credentials_are_wrong() -> None:
     manager.check_credentials.return_value = False
 
     with pytest.raises(BenchError, match="db.example.com"):
-        _command()._provision_or_verify(manager, "MariaDB")
+        _initializer()._provision_or_verify(manager, "MariaDB")
 
     manager.provision.assert_not_called()
