+++ tests/pilot/commands/test_setup_production.py
@@ -1,4 +1,4 @@
-"""Tests for SetupProductionCommand helpers and letsencrypt gating.
+"""Tests for ProductionSetup helpers and letsencrypt gating.
 
 The full `run()` touches sudo/systemd, so these exercise the pure helpers:
 admin-domain handling, in-place toml persistence, and the cert-needed check.
@@ -10,7 +10,7 @@ from pathlib import Path
 
 import pytest
 
-from pilot.commands.setup.production import SetupProductionCommand
+from pilot.core.production_setup import ProductionSetup
 from pilot.config.bench import BenchConfig
 from pilot.core.bench import Bench
 from pilot.exceptions import BenchError
@@ -37,7 +37,7 @@ def _make_bench(tmp_path: Path, name: str = "prod", *, admin_domain: str = "prod
 
 def test_persist_preserves_other_fields(tmp_path: Path) -> None:
     bench = _make_bench(tmp_path)
-    cmd = SetupProductionCommand(bench)
+    cmd = ProductionSetup(bench)
     cmd._persist({"admin": {"domain": "admin.example.com"}})
 
     data = tomllib.loads((bench.path / "bench.toml").read_text())
@@ -50,7 +50,7 @@ def test_persist_preserves_other_fields(tmp_path: Path) -> None:
 
 def test_check_admin_domain_uses_toml_value(tmp_path: Path) -> None:
     bench = _make_bench(tmp_path, admin_domain="keep.example.com")
-    cmd = SetupProductionCommand(bench)
+    cmd = ProductionSetup(bench)
     cmd._check_admin_domain()  # must not prompt or raise
     assert bench.config.admin.domain == "keep.example.com"
 
@@ -58,7 +58,7 @@ def test_check_admin_domain_uses_toml_value(tmp_path: Path) -> None:
 def test_check_admin_domain_rejects_sibling_owned(tmp_path: Path) -> None:
     _make_bench(tmp_path, name="other", admin_domain="shared.example.com")
     bench = _make_bench(tmp_path, name="prod", admin_domain="shared.example.com")
-    cmd = SetupProductionCommand(bench)
+    cmd = ProductionSetup(bench)
     with pytest.raises(BenchError, match="already used by bench 'other'"):
         cmd._check_admin_domain()
 
@@ -68,9 +68,9 @@ def test_check_admin_domain_grandfathers_existing_non_matching(tmp_path: Path, m
 
     monkeypatch.setattr(DomainRouteProvider, "wildcard_domains", staticmethod(lambda: ["*.node1.example.com"]))
     bench = _make_bench(tmp_path, admin_domain="node1.example.com")  # apex, can't match wildcard
-    SetupProductionCommand(bench)._check_admin_domain()  # existing -> no raise
+    ProductionSetup(bench)._check_admin_domain()  # existing -> no raise
 
-    cmd = SetupProductionCommand(bench, admin_domain="other.example.com")
+    cmd = ProductionSetup(bench, admin_domain="other.example.com")
     cmd._resolve_target()
     with pytest.raises(BenchError, match="must match one of this bench's wildcard"):
         cmd._check_admin_domain()
@@ -94,7 +94,7 @@ def test_needs_letsencrypt(tmp_path: Path) -> None:
 
 def test_resolve_target_uses_flag_over_config(tmp_path: Path) -> None:
     bench = _make_bench(tmp_path, process_manager="supervisor")
-    cmd = SetupProductionCommand(bench, process_manager="systemd")
+    cmd = ProductionSetup(bench, process_manager="systemd")
     cmd._resolve_target()
     assert bench.config.production.process_manager == "systemd"
     assert bench.config.production.enabled is True
@@ -104,21 +104,21 @@ def test_resolve_target_uses_flag_over_config(tmp_path: Path) -> None:
 
 def test_resolve_target_applies_tls_flag(tmp_path: Path) -> None:
     bench = _make_bench(tmp_path, tls=True)
-    cmd = SetupProductionCommand(bench, process_manager="systemd", admin_tls=False)
+    cmd = ProductionSetup(bench, process_manager="systemd", admin_tls=False)
     cmd._resolve_target()
     assert bench.config.admin.tls is False
 
 
 def test_resolve_target_normalizes_supervisord(tmp_path: Path) -> None:
     bench = _make_bench(tmp_path, process_manager="supervisor")
-    cmd = SetupProductionCommand(bench, process_manager="supervisord")
+    cmd = ProductionSetup(bench, process_manager="supervisord")
     cmd._resolve_target()
     assert bench.config.production.process_manager == "supervisor"
 
 
 def test_resolve_target_defaults_to_systemd(tmp_path: Path) -> None:
     bench = _make_bench(tmp_path, process_manager="none")
-    cmd = SetupProductionCommand(bench)
+    cmd = ProductionSetup(bench)
     cmd._resolve_target()
     assert bench.config.production.process_manager == "systemd"
     assert bench.config.production.enabled is True
@@ -126,14 +126,14 @@ def test_resolve_target_defaults_to_systemd(tmp_path: Path) -> None:
 
 def test_resolve_target_applies_admin_domain(tmp_path: Path) -> None:
     bench = _make_bench(tmp_path)
-    cmd = SetupProductionCommand(bench, process_manager="systemd", admin_domain="admin-new.example.com")
+    cmd = ProductionSetup(bench, process_manager="systemd", admin_domain="admin-new.example.com")
     cmd._resolve_target()
     assert bench.config.admin.domain == "admin-new.example.com"
 
 
 def test_resolve_target_applies_letsencrypt_email(tmp_path: Path) -> None:
     bench = _make_bench(tmp_path)
-    cmd = SetupProductionCommand(bench, process_manager="systemd", letsencrypt_email="me@example.com")
+    cmd = ProductionSetup(bench, process_manager="systemd", letsencrypt_email="me@example.com")
     cmd._resolve_target()
     assert bench.config.letsencrypt.email == "me@example.com"
 
@@ -141,7 +141,7 @@ def test_resolve_target_applies_letsencrypt_email(tmp_path: Path) -> None:
 def test_require_production_inputs_needs_admin_domain(tmp_path: Path) -> None:
     # Fresh, undeployed bench: empty domain, no process manager yet (so it loads).
     bench = _make_bench(tmp_path, admin_domain="", process_manager="")
-    cmd = SetupProductionCommand(bench, process_manager="systemd")
+    cmd = ProductionSetup(bench, process_manager="systemd")
     cmd._resolve_target()
     with pytest.raises(BenchError, match="admin domain is required"):
         cmd._require_production_inputs()
@@ -149,7 +149,7 @@ def test_require_production_inputs_needs_admin_domain(tmp_path: Path) -> None:
 
 def test_require_production_inputs_needs_email_for_tls(tmp_path: Path) -> None:
     bench = _make_bench(tmp_path, admin_domain="admin.example.com", tls=True, email="")
-    cmd = SetupProductionCommand(bench, process_manager="systemd")
+    cmd = ProductionSetup(bench, process_manager="systemd")
     cmd._resolve_target()
     with pytest.raises(BenchError, match="contact email is required"):
         cmd._require_production_inputs()
@@ -157,7 +157,7 @@ def test_require_production_inputs_needs_email_for_tls(tmp_path: Path) -> None:
 
 def test_require_production_inputs_passes_with_domain_and_email(tmp_path: Path) -> None:
     bench = _make_bench(tmp_path, admin_domain="admin.example.com", tls=True, email="me@example.com")
-    cmd = SetupProductionCommand(bench, process_manager="systemd")
+    cmd = ProductionSetup(bench, process_manager="systemd")
     cmd._resolve_target()
     cmd._require_production_inputs()  # no raise
 
@@ -179,7 +179,7 @@ def test_setup_monitoring_persists_log_path_to_toml(tmp_path: Path, monkeypatch)
 
     bench = _make_bench(tmp_path, process_manager="systemd")
     bench.config.production.enabled = True
-    cmd = SetupProductionCommand(bench)
+    cmd = ProductionSetup(bench)
 
     monkeypatch.setattr(MonitorConfigurator, "install", lambda self: None)
 
@@ -195,7 +195,7 @@ def test_setup_monitoring_log_path_is_path_on_config(tmp_path: Path, monkeypatch
 
     bench = _make_bench(tmp_path, process_manager="systemd")
     bench.config.production.enabled = True
-    cmd = SetupProductionCommand(bench)
+    cmd = ProductionSetup(bench)
 
     monkeypatch.setattr(MonitorConfigurator, "install", lambda self: None)
 
@@ -209,12 +209,12 @@ def test_setup_monitoring_log_path_is_path_on_config(tmp_path: Path, monkeypatch
 
 
 def test_setup_letsencrypt_reraises_by_default(tmp_path: Path, monkeypatch) -> None:
-    from pilot.commands.setup import letsencrypt as letsencrypt_module
+    from pilot.core.bench import Bench
 
     bench = _make_bench(tmp_path, admin_domain="admin.example.com", email="x@y.com")
-    monkeypatch.setattr(letsencrypt_module.SetupLetsEncryptCommand, "run",
+    monkeypatch.setattr(Bench, "setup_letsencrypt",
                          lambda self: (_ for _ in ()).throw(RuntimeError("dns not ready")))
-    cmd = SetupProductionCommand(bench)
+    cmd = ProductionSetup(bench)
 
     with pytest.raises(RuntimeError, match="dns not ready"):
         cmd._setup_letsencrypt_if_needed()
@@ -225,12 +225,12 @@ def test_setup_letsencrypt_swallows_when_best_effort(tmp_path: Path, monkeypatch
     shouldn't roll back an otherwise-working deployment just because a cert
     can't issue yet - e.g. DNS for a domain created moments ago hasn't
     propagated. The bench should stay live on HTTP instead."""
-    from pilot.commands.setup import letsencrypt as letsencrypt_module
+    from pilot.core.bench import Bench
 
     bench = _make_bench(tmp_path, admin_domain="admin.example.com", email="x@y.com")
-    monkeypatch.setattr(letsencrypt_module.SetupLetsEncryptCommand, "run",
+    monkeypatch.setattr(Bench, "setup_letsencrypt",
                          lambda self: (_ for _ in ()).throw(RuntimeError("dns not ready")))
-    cmd = SetupProductionCommand(bench, best_effort_tls=True)
+    cmd = ProductionSetup(bench, best_effort_tls=True)
 
     cmd._setup_letsencrypt_if_needed()  # must not raise
 
@@ -245,7 +245,7 @@ def test_persist_production_state_writes_enabled_and_drops_nginx(tmp_path: Path)
         '[production]\nprocess_manager = "supervisor"\n',
         '[production]\nprocess_manager = "supervisor"\nnginx = true\n',
     ))
-    cmd = SetupProductionCommand(bench, process_manager="systemd")
+    cmd = ProductionSetup(bench, process_manager="systemd")
     cmd._resolve_target()
     cmd._persist_production_state()
     data = tomllib.loads(toml_path.read_text())
